- corrected the prototype of count-in-range to take integer counts, not double
git-svn-id: svn://svn.code.sf.net/p/gwyddion/code@10910 25697fe4-d644-49b8-9d30-be5465f2adf3

+++ branches/GWYDDION-3/libgwy/field-statistics.h
@@ -97,8 +97,8 @@ guint   gwy_field_part_count_in_range(const GwyField *field,
                                       gdouble lower,
                                       gdouble upper,
                                       gboolean strict,
-                                      gdouble *nabove,
-                                      gdouble *nbelow);
+                                      guint *nabove,
+                                      guint *nbelow);
 
 G_END_DECLS
